+++ nncf/experimental/torch/nas/bootstrapNAS/__init__.py
@@ -10,7 +10,3 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 """
-
-from nncf.experimental.torch.nas.bootstrapNAS.training.training_algorithm import EpochBasedTrainingAlgorithm
-from nncf.experimental.torch.nas.bootstrapNAS.elasticity import elasticity_builder as elasticity_algo
-from nncf.experimental.torch.nas.bootstrapNAS.training import progressive_shrinking_builder as ps_algo
