+++ nncf/experimental/torch/nas/bootstrapNAS/training/model_creator_helpers.py
@@ -33,11 +33,11 @@ def create_compressed_model_from_algo_names(nncf_network: NNCFNetwork,
                                             config: NNCFConfig,
                                             algo_names: List[str],
                                             dump_graphs: bool = True) -> Tuple[BaseController, NNCFNetwork]:
-    set_debug_log_dir(config.get("log_dir", "."))
+    set_debug_log_dir(config.get("log_dir", ""))
 
     if dump_graphs:
         original_model_graph = nncf_network.get_original_graph()
-        original_model_graph.visualize_graph(osp.join(config.get("log_dir", "."), "original_graph.dot"))
+        original_model_graph.visualize_graph(osp.join(config.get("log_dir", ""), "original_graph.dot"))
 
     builder = create_compression_algorithm_builder_from_algo_names(algo_names, config, should_init=True)
 
@@ -50,7 +50,7 @@ def create_compressed_model_from_algo_names(nncf_network: NNCFNetwork,
 
     if dump_graphs and is_main_process():
         original_model_graph = compressed_model.get_graph()
-        original_model_graph.visualize_graph(osp.join(config.get("log_dir", "."), "compressed_graph.dot"))
+        original_model_graph.visualize_graph(osp.join(config.get("log_dir", ""), "compressed_graph.dot"))
 
     synchronize_all_processes_in_distributed_mode()
     return compression_ctrl, compressed_model
