+++ nncf/experimental/torch/nas/bootstrapNAS/training/progressive_shrinking_controller.py
@@ -21,12 +21,12 @@ from nncf.common.initialization.batchnorm_adaptation import BatchnormAdaptationA
 from nncf.common.statistics import NNCFStatistics
 from nncf.common.utils.logger import logger as nncf_logger
 from nncf.torch.algo_selector import ZeroCompressionLoss
-from nncf.torch.nas.bootstrapNAS.elasticity.elasticity_controller import ElasticityController
-from nncf.torch.nas.bootstrapNAS.elasticity.elasticity_dim import ElasticityDim
-from nncf.torch.nas.bootstrapNAS.elasticity.multi_elasticity_handler import MultiElasticityHandler
-from nncf.torch.nas.bootstrapNAS.training.base_training import BNASTrainingController
-from nncf.torch.nas.bootstrapNAS.training.scheduler import BootstrapNASScheduler
-from nncf.torch.nas.bootstrapNAS.training.stage_descriptor import StageDescriptor
+from nncf.experimental.torch.nas.bootstrapNAS.elasticity.elasticity_controller import ElasticityController
+from nncf.experimental.torch.nas.bootstrapNAS.elasticity.elasticity_dim import ElasticityDim
+from nncf.experimental.torch.nas.bootstrapNAS.elasticity.multi_elasticity_handler import MultiElasticityHandler
+from nncf.experimental.torch.nas.bootstrapNAS.training.base_training import BNASTrainingController
+from nncf.experimental.torch.nas.bootstrapNAS.training.scheduler import BootstrapNASScheduler
+from nncf.experimental.torch.nas.bootstrapNAS.training.stage_descriptor import StageDescriptor
 from nncf.torch.nncf_network import NNCFNetwork
